requirements.txt
flask>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.196
torch>=2.0.0
torchvision>=0.15.0
flask-socketio>=5.3.0
python-engineio>=4.5.1
python-socketio>=5.8.0
eventlet>=0.33.3
gunicorn>=21.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.1
